Extended page header with useful content
As we want the title to be defined in content, I've added a variable for
this.

Also site description is an important thing for Google and user, so this
can be defined through description on every site.

Also some icons are important to be recognized. They are optionally.
Remove the comments and place them in the public/ folder of your site to
be linked.